Safari-609.1.11.1/LayoutTests/http/tests/security/window-onerror-exception-in-iframe-expected.txt

CONSOLE MESSAGE: line 7: IframeException
Test that exceptions in iframe are not reported to the main frame window.onerror handler. window.onerror should print exactly one line.

Main frame window.onerror: MainFrameException at window-onerror-exception-in-iframe.html:34:27 MainFrameException
DONE
